Added header component to calendar and dashboard
