+++ src/lib/components/Header.svelte
@@ -1,10 +1,75 @@
 <script lang="ts">
+	import { WeekDay } from "$lib/types";
+
+	const currentDate = new Date();
+	const dayNumber = currentDate.getDay();
+	const weekdays = Object.values(WeekDay);
+	const DayOfWeek = weekdays[dayNumber];
+
+	const month = currentDate.toLocaleString("en-US", {
+		month: "long"
+	});
+
+	const day = currentDate.toLocaleString("en-US", {
+		day: "numeric"
+	});
+
+	const year = currentDate.toLocaleString("en-US", {
+		year: "numeric"
+	});
+
+	function getNumSuffix(i: number): string {
+		let selected;
+
+		if ((i > 3 && i < 21) || i % 10 > 3) {
+			selected = 0;
+		} else {
+			selected = i % 10;
+		}
+
+		switch (selected) {
+			case 0:
+				return "th";
+			case 1:
+				return "st";
+			case 2:
+				return "nd";
+			default:
+				return "rd";
+		}
+	}
+
+	const dateSuffix = getNumSuffix(currentDate.getDate());
 	
+	export let layout: string;
 </script>
 
 <div>
     <div class="px-12 py-2">
-        <slot></slot>
+		{#if layout == 'dashboard'}
+		<div class="header-component">
+			<div class="flex flex-row w-full">
+				<div class="flex justify-between w-full">
+					<h1 class="main-text header capitalize">
+						{DayOfWeek}, {month}
+						{day}<sup class="text-primary">{dateSuffix}</sup>
+					</h1>
+					<h1 class="main-text header-year text-primary px-8">{year}</h1>
+				</div>
+			</div>
+		</div>
+		{:else if layout === 'calendar'}
+		<div class="flex flex-row justify-between">
+			
+			<div class="flex flex-col items-start">
+				<h1 class="header-year font-bold font-kaisei">
+					<span>{month}</span> <span class="text-primary">{year}</span>
+				</h1>
+				<h2 class="subheading font-bold font-kaisei text-primary text-xl">{month}/{day}</h2>
+			</div>
+			<slot></slot>
+		</div>
+		{/if}
         <div class="border-t-4 border-solid border-gray-500 my-4 me-36"></div>
     </div>
 </div>
