+++ src/routes/dashboard/+page.svelte
@@ -1,64 +1,11 @@
 <script lang="ts">
 	import BookingForm from "$lib/components/dashboard/BookingForm.svelte";
 	import OfficeHourSummary from "$lib/components/dashboard/OfficeHourSummary.svelte";
-	import { WeekDay } from "$lib/types";
 	import Header from "$lib/components/Header.svelte";
-
-	const currentDate = new Date();
-	const dayNumber = currentDate.getDay();
-	const weekdays = Object.values(WeekDay);
-	const DayOfWeek = weekdays[dayNumber];
-
-	const month = currentDate.toLocaleString("en-US", {
-		month: "long"
-	});
-
-	const day = currentDate.toLocaleString("en-US", {
-		day: "numeric"
-	});
-
-	const year = currentDate.toLocaleString("en-US", {
-		year: "numeric"
-	});
-
-	function getNumSuffix(i: number): string {
-		let selected;
-
-		if ((i > 3 && i < 21) || i % 10 > 3) {
-			selected = 0;
-		} else {
-			selected = i % 10;
-		}
-
-		switch (selected) {
-			case 0:
-				return "th";
-			case 1:
-				return "st";
-			case 2:
-				return "nd";
-			default:
-				return "rd";
-		}
-	}
-
-	const dateSuffix = getNumSuffix(currentDate.getDate());
 </script>
 
 <div class="px-12 py-2">
-	<Header>
-		<div class="header-component">
-			<div class="flex flex-row w-full">
-				<div class="flex justify-between w-full">
-					<h1 class="main-text header capitalize">
-						{DayOfWeek}, {month}
-						{day}<sup class="text-primary">{dateSuffix}</sup>
-					</h1>
-					<h1 class="main-text header-year text-primary px-8">{year}</h1>
-				</div>
-			</div>
-		</div>
-	</Header>
+	<Header layout="dashboard" />
 
 	<div class="px-12 py-6 mx-auto flex flex-row justify-center">
 		<div class="appointment my-8 mx-auto">
